chore: use `Modal` in the `DebuggerModal`
