+++ src/components/QuickReferenceModal.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
-import { faUpRightFromSquare, faXmark } from '@fortawesome/free-solid-svg-icons'
-import Fa from 'svelte-fa'
+import { faUpRightFromSquare } from '@fortawesome/free-solid-svg-icons'
 import Modal from './Modal.svelte'
 import type { ReferenceDoc } from './types'
 
@@ -12,21 +11,19 @@ interface Props {
 const { referenceDoc, onClose }: Props = $props()
 
 function openExternalLink() {
-  // onClose()
   window.open(`${referenceDoc.url}#${referenceDoc.anchor}`, '_blank')
 }
+
+const actions = $derived([
+  {
+    title: 'Open the documentation in a separate page',
+    onclick: openExternalLink,
+    icon: faUpRightFromSquare
+  }
+])
 </script>
 
-<Modal {onClose}>
-  <div class="header">
-    <div class="title">Quick reference</div>
-    <button type="button" title="Open the documentation in a separate page" onclick={openExternalLink}>
-      <Fa icon={faUpRightFromSquare} size="lg" />
-    </button>
-    <button type="button" title="Close (Esc)" class="close" onclick={onClose}>
-      <Fa icon={faXmark} size="lg" />
-    </button>
-  </div>
+<Modal title="Quick reference" {actions} {onClose}>
   <div class="content">
     {#if referenceDoc.doc}
       {@html referenceDoc.doc}
@@ -37,38 +34,7 @@ function openExternalLink() {
 </Modal>
 
 <style>
-  .header {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    background: var(--theme-color);
-
-    .title {
-      flex: 1;
-      font-weight: bold;
-      padding: var(--padding);
-    }
-
-    button {
-      width: 40px;
-      align-self: stretch;
-      background: transparent;
-      border: none;
-      color: var(--color);
-      cursor: pointer;
-    }
-
-    button:hover {
-      background: var(--theme-color-highlight);
-    }
-
-    button.close:hover {
-      color: var(--error-color);
-    }
-  }
-
   .content {
-    position: relative;
     margin: calc(2 * var(--padding));
   }
