+++ src/layouts/Layout.astro
@@ -50,6 +50,7 @@ const { title, description } = Astro.props
 		--button-margin: 5px;
 		--button-border: 1px solid #bfbfbf;
 		--border-radius: 3px;
+		--input-border: 1px solid #bfbfbf;
 		--tab-border-width: 5px;
 		--table-cell-border: 1px solid #e5e5e5;
 		--table-cell-padding: 5px 10px;
